In principle, dynamic scripts should be shown in two ways in MD context menus: - shown as enabled/disabled action - only show if enabled
The first approach results in cluttered menus; potentially with lots of disabled dynamic script actions
with only a few enabled dynamic script actions.

The second approach results in minimially cluttered menus because all the dynamic script actions shown are enabled, by definition.
However, it can also result in confusion if the user expected to see a particular dynamic script action enabled but it is not and thus not shown.
This update addresses the confusion:
- on the first pass where a dynamic script is not enabled, an explanation will be shown in the MD log & MD gui.
  Subsequent invocations of the same context menu will not show the same explanation.
- if the dynamic script specification becomes enabled somehow, then an explanation will be shown in the MD log & MD gui as well.

THis is intended to help troubleshoot problems in the configuration of a dynamic script that can prevent its intended enablement behavior.